+++ packages/react-dom/src/Link.tsx
@@ -25,15 +25,28 @@ export type AsyncLinkProps = BaseLinkProps &
 
 export const Link = React.forwardRef(
   (props: LinkProps, ref: React.Ref<any>) => {
-    const url = useURL(props);
+    const {
+      // url
+      name,
+      params,
+      query,
+      hash,
+      // navigation
+      state,
+      onNav,
+      method,
+      // props
+      children,
+      anchor: Anchor = "a",
+      ...rest
+    } = props;
+    const url = useURL({ name, params, query, hash });
     const { eventHandler } = useNavigationHandler<
       React.MouseEvent<HTMLElement>
-    >(url, props, canNavigate);
-
-    const { anchor: Anchor = "a", forward, children } = props;
+    >({ url, state, onNav, method, canNavigate, target: rest.target });
 
     return (
-      <Anchor {...forward} onClick={eventHandler} href={url} ref={ref}>
+      <Anchor {...rest} onClick={eventHandler} href={url} ref={ref}>
         {children}
       </Anchor>
     );
@@ -42,15 +55,28 @@ export const Link = React.forwardRef(
 
 export const AsyncLink = React.forwardRef(
   (props: AsyncLinkProps, ref: React.Ref<any>) => {
-    const url = useURL(props);
+    const {
+      // url
+      name,
+      params,
+      query,
+      hash,
+      // navigation
+      state,
+      onNav,
+      method,
+      // props
+      children,
+      anchor: Anchor = "a",
+      ...rest
+    } = props;
+    const url = useURL({ name, params, query, hash });
     const { eventHandler, navigating } = useStatefulNavigationHandler<
       React.MouseEvent<HTMLElement>
-    >(url, props, canNavigate);
-
-    const { anchor: Anchor = "a", forward, children } = props;
+    >({ url, state, onNav, method, canNavigate, target: rest.target });
 
     return (
-      <Anchor {...forward} onClick={eventHandler} href={url} ref={ref}>
+      <Anchor {...rest} onClick={eventHandler} href={url} ref={ref}>
         {children(navigating)}
       </Anchor>
     );