+++ packages/react-dom/src/utils.ts
@@ -3,8 +3,8 @@ import { MouseEvent } from "react";
 export function canNavigate(event: MouseEvent<HTMLElement>, target?: string) {
   return (
     !event.defaultPrevented &&
+    !target &&
     event.button === 0 &&
-    !(event.metaKey || event.altKey || event.ctrlKey || event.shiftKey) &&
-    (!target || target === "_self")
+    !(event.metaKey || event.altKey || event.ctrlKey || event.shiftKey)
   );
 }