+++ packages/react-universal/src/hooks/useNavigationHandler.ts
@@ -5,10 +5,12 @@ import { ReactNode } from "react";
 import { NavType } from "@hickory/root";
 
 export interface NavigationHookProps<T> {
+  url: string;
   onNav?: (e: T) => void;
   method?: NavType;
-  forward?: object;
+  target?: string;
   state?: any;
+  canNavigate: CanNavigate<T>;
 }
 export type NavigatingChildren = (navigating: boolean) => ReactNode;
 export type CanNavigate<T> = (e: T, target?: string) => boolean;
@@ -18,30 +20,29 @@ function defaultCanNavigate() {
 }
 
 export function useNavigationHandler<T extends React.BaseSyntheticEvent>(
-  url: string,
-  props: NavigationHookProps<T>,
-  canNavigate: CanNavigate<T> = defaultCanNavigate
+  props: NavigationHookProps<T>
 ) {
+  const {
+    url,
+    onNav,
+    method,
+    target,
+    state,
+    canNavigate = defaultCanNavigate
+  } = props;
   const router = useRouter();
-  // @ts-ignore
-  const target = props.forward && props.forward.target;
   const eventHandler = React.useCallback(
     function eventHandler(event: T) {
-      if (props.onNav) {
-        props.onNav(event);
+      if (onNav) {
+        onNav(event);
       }
 
       if (canNavigate(event, target)) {
         event.preventDefault();
-
-        router.navigate({
-          url,
-          state: props.state,
-          method: props.method
-        });
+        router.navigate({ url, state, method });
       }
     },
-    [url, props.method, props.state, props.onNav, target]
+    [url, method, state, onNav, target]
   );
   return {
     url,
@@ -51,11 +52,15 @@ export function useNavigationHandler<T extends React.BaseSyntheticEvent>(
 
 export function useStatefulNavigationHandler<
   T extends React.BaseSyntheticEvent
->(
-  url: string,
-  props: NavigationHookProps<T>,
-  canNavigate: CanNavigate<T> = defaultCanNavigate
-) {
+>(props: NavigationHookProps<T>) {
+  const {
+    url,
+    onNav,
+    method,
+    target,
+    state,
+    canNavigate = defaultCanNavigate
+  } = props;
   const router = useRouter();
   const removeCallbacks = React.useRef(undefined);
   const [navigating, setNavigating] = React.useState(false);
@@ -68,33 +73,29 @@ export function useStatefulNavigationHandler<
     };
   }, []);
 
-  // @ts-ignore
-  const target = props.forward && props.forward.target;
   const eventHandler = React.useCallback(
     function eventHandler(event: T) {
-      if (props.onNav) {
-        props.onNav(event);
+      if (onNav) {
+        onNav(event);
       }
 
       if (canNavigate(event, target)) {
         event.preventDefault();
-
         const done = () => {
           removeCallbacks.current = undefined;
           setNavigating(false);
         };
         setNavigating(true);
-
         removeCallbacks.current = router.navigate({
           url,
-          state: props.state,
-          method: props.method,
+          state,
+          method,
           cancelled: done,
           finished: done
         });
       }
     },
-    [url, props.method, props.state, props.onNav, target]
+    [url, method, state, onNav, target]
   );
 
   return {