+++ packages/react-universal/types/hooks/useNavigationHandler.d.ts
@@ -2,18 +2,20 @@ import React from "react";
 import { ReactNode } from "react";
 import { NavType } from "@hickory/root";
 export interface NavigationHookProps<T> {
+    url: string;
     onNav?: (e: T) => void;
     method?: NavType;
-    forward?: object;
+    target?: string;
     state?: any;
+    canNavigate: CanNavigate<T>;
 }
 export declare type NavigatingChildren = (navigating: boolean) => ReactNode;
 export declare type CanNavigate<T> = (e: T, target?: string) => boolean;
-export declare function useNavigationHandler<T extends React.BaseSyntheticEvent>(url: string, props: NavigationHookProps<T>, canNavigate?: CanNavigate<T>): {
+export declare function useNavigationHandler<T extends React.BaseSyntheticEvent>(props: NavigationHookProps<T>): {
     url: string;
     eventHandler: (event: T) => void;
 };
-export declare function useStatefulNavigationHandler<T extends React.BaseSyntheticEvent>(url: string, props: NavigationHookProps<T>, canNavigate?: CanNavigate<T>): {
+export declare function useStatefulNavigationHandler<T extends React.BaseSyntheticEvent>(props: NavigationHookProps<T>): {
     url: string;
     eventHandler: (event: T) => void;
     navigating: boolean;