+++ website/src/components/links/ActiveLink.js
@@ -2,7 +2,7 @@ import React from "react";
 import { useActive, Link } from "@curi/react-dom";
 
 const ActiveLink = React.forwardRef(
-  ({ name, params, forward, ...rest }, ref) => {
+  ({ name, params, className = "", ...rest }, ref) => {
     const active = useActive({ name, params, partial: true });
     return (
       <Link
@@ -10,13 +10,7 @@ const ActiveLink = React.forwardRef(
         name={name}
         params={params}
         {...rest}
-        forward={{
-          ...forward,
-          className: [
-            (forward && forward.className) || "",
-            active ? "active" : ""
-          ].join(" ")
-        }}
+        className={`${className} ${active ? "active" : ""}`}
       />
     );
   }